IGNITE-16185 Improve the interface of UserObjectMarshaller
- move application of special serialization methods from ClassDescriptor
