+++ modules/network/src/main/java/org/apache/ignite/internal/network/serialization/SpecialSerializationMethods.java
@@ -17,89 +17,10 @@
 
 package org.apache.ignite.internal.network.serialization;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
-import java.lang.reflect.Method;
-import java.util.Objects;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
 /**
  * Encapsulates special serialization methods like writeReplace()/readResolve() for convenient invocation.
  */
-class SpecialSerializationMethods {
-    /** MethodHandle that can be used to invoke writeReplace() on the target class. */
-    @Nullable
-    private final MethodHandle writeReplaceHandle;
-
-    /** MethodHandle that can be used to invoke readResolve() on the target class. */
-    @Nullable
-    private final MethodHandle readResolveHandle;
-
-    /**
-     * Creates a new instance from the provided descriptor.
-     *
-     * @param descriptor class descriptor on which class to operate
-     */
-    public SpecialSerializationMethods(ClassDescriptor descriptor) {
-        writeReplaceHandle = descriptor.hasWriteReplace() ? writeReplaceHandle(descriptor) : null;
-        readResolveHandle = descriptor.hasReadResolve() ? readResolveHandle(descriptor) : null;
-    }
-
-    private static MethodHandle writeReplaceHandle(ClassDescriptor descriptor) {
-        Method writeReplaceMethod = findWriteReplaceMethod(descriptor);
-
-        try {
-            return MethodHandles.lookup()
-                        .unreflect(writeReplaceMethod)
-                        .asType(MethodType.methodType(Object.class, Object.class));
-        } catch (IllegalAccessException e) {
-            throw new ReflectionException("writeReplace() cannot be unreflected", e);
-        }
-    }
-
-    private static Method findWriteReplaceMethod(ClassDescriptor descriptor) {
-        Method writeReplaceMethod;
-        try {
-            writeReplaceMethod = descriptor.clazz().getDeclaredMethod("writeReplace");
-        } catch (NoSuchMethodException e) {
-            throw new ReflectionException("writeReplace() was not found on " + descriptor.clazz()
-                    + " even though the descriptor says the class has the method", e);
-        }
-
-        writeReplaceMethod.setAccessible(true);
-
-        return writeReplaceMethod;
-    }
-
-    private static MethodHandle readResolveHandle(ClassDescriptor descriptor) {
-        Method readResolveMethod = findReadResolveMethod(descriptor);
-
-        try {
-            return MethodHandles.lookup()
-                    .unreflect(readResolveMethod)
-                    .asType(MethodType.methodType(Object.class, Object.class));
-        } catch (IllegalAccessException e) {
-            throw new ReflectionException("readResolve() cannot be unreflected", e);
-        }
-    }
-
-    @NotNull
-    private static Method findReadResolveMethod(ClassDescriptor descriptor) {
-        Method readResolveMethod;
-        try {
-            readResolveMethod = descriptor.clazz().getDeclaredMethod("readResolve");
-        } catch (NoSuchMethodException e) {
-            throw new ReflectionException("readResolve() was not found on " + descriptor.clazz()
-                    + " even though the descriptor says the class has the method", e);
-        }
-
-        readResolveMethod.setAccessible(true);
-
-        return readResolveMethod;
-    }
-
+public interface SpecialSerializationMethods {
     /**
      * Invokes {@code writeReplace()} on the target object. Should only be used if the target descriptor supports
      * the method (that is, its serialization type is
@@ -112,17 +33,7 @@ class SpecialSerializationMethods {
      * @return invocation result
      * @throws SpecialMethodInvocationException if the invocation fails
      */
-    public Object writeReplace(Object object) throws SpecialMethodInvocationException {
-        Objects.requireNonNull(writeReplaceHandle);
-
-        try {
-            return writeReplaceHandle.invokeExact(object);
-        } catch (Error e) {
-            throw e;
-        } catch (Throwable e) {
-            throw new SpecialMethodInvocationException("writeReplace() invocation failed on " + object, e);
-        }
-    }
+    Object writeReplace(Object object) throws SpecialMethodInvocationException;
 
     /**
      * Invokes {@code readResolve()} on the target object. Should only be used if the target descriptor supports
@@ -136,15 +47,5 @@ class SpecialSerializationMethods {
      * @return invocation result
      * @throws SpecialMethodInvocationException if the invocation fails
      */
-    public Object readResolve(Object object) throws SpecialMethodInvocationException {
-        Objects.requireNonNull(readResolveHandle);
-
-        try {
-            return readResolveHandle.invokeExact(object);
-        } catch (Error e) {
-            throw e;
-        } catch (Throwable e) {
-            throw new SpecialMethodInvocationException("readResolve() invocation failed on " + object, e);
-        }
-    }
+    Object readResolve(Object object) throws SpecialMethodInvocationException;
 }
